replace files and organize
