+++ src/controllers/MenuController.java
@@ -1,4 +1,4 @@
-package application;
+package controllers;
 
 import java.net.URL;
 import java.util.ResourceBundle;
